compile

+++ femlib/src/EMeshStiffnessSparse.cpp
@@ -4,7 +4,7 @@
 void ElementMesh::getStiffnessSparse(std::vector<int> & I, std::vector<int> & J, 
                                      std::vector<float> &val)
 {
-  int N = 3* x.size();
+  int N = 3* (int)x.size();
   Eigen::SparseMatrix<float> Ksparse(N,N);
   for(unsigned int ii = 0;ii<e.size();ii++){
     MatrixXd K  = getStiffness(ii);
